+++ Sveti/App/Screens/Diary/DiaryVM.swift
@@ -55,8 +55,8 @@ class DiaryVM {
     guard let notes = notes else { return nil }
     guard notes.count > 1 else { return nil }
     let mathHelper = MathHelper()
-    var totalScore: Float = 0
-    var totalNotes: Float = 0
+    var totalScore: Double = 0
+    var totalNotes: Double = 0
 
     for note in notes {
       guard let average = note.mood?.average else { continue }
